requirements.txt
Flask==2.1.0
flask_sqlalchemy
mysql-connector-python
werkzeug==2.0.2
sqlalchemy==1.4.49
PyJWT==2.3.0 
twilio

